feat: refactor job module to include manifest module and update related imports
