+++ packages/apps/job-launcher/server/jest.config.ts
@@ -1,13 +1,15 @@
+import { createDefaultPreset } from 'ts-jest';
+
+const jestTsPreset = createDefaultPreset({});
+
 module.exports = {
+  ...jestTsPreset,
   coverageDirectory: '../coverage',
   collectCoverageFrom: ['**/*.(t|j)s'],
   moduleFileExtensions: ['js', 'json', 'ts'],
   rootDir: 'src',
   testEnvironment: 'node',
   testRegex: '.*\\.spec\\.ts$',
-  transform: {
-    '^.+\\.(t|j)s$': 'ts-jest',
-  },
   moduleNameMapper: {
     '^uuid$': require.resolve('uuid'),
     '^typeorm$': require.resolve('typeorm'),
