+++ packages/apps/job-launcher/server/src/modules/job/job.service.ts
@@ -1,16 +1,18 @@
 /* eslint-disable @typescript-eslint/no-non-null-assertion */
 import {
+  HMToken,
+  HMToken__factory,
+} from '@human-protocol/core/typechain-types';
+import {
   ChainId,
   EscrowClient,
   EscrowStatus,
   EscrowUtils,
-  NETWORKS,
-  StorageParams,
-  Encryption,
   KVStoreKeys,
   KVStoreUtils,
+  NETWORKS,
+  StorageParams,
 } from '@human-protocol/sdk';
-import { v4 as uuidv4 } from 'uuid';
 import {
   HttpStatus,
   Inject,
@@ -18,109 +20,65 @@ import {
   Logger,
   ValidationError,
 } from '@nestjs/common';
+import { ModuleRef } from '@nestjs/core';
 import { validate } from 'class-validator';
+import Decimal from 'decimal.js';
 import { ethers } from 'ethers';
-import { AuthConfigService } from '../../common/config/auth-config.service';
 import { ServerConfigService } from '../../common/config/server-config.service';
-import { CvatConfigService } from '../../common/config/cvat-config.service';
-import { PGPConfigService } from '../../common/config/pgp-config.service';
-import { Web3ConfigService } from '../../common/config/web3-config.service';
+import { CANCEL_JOB_STATUSES } from '../../common/constants';
 import {
-  ErrorBucket,
   ErrorEscrow,
   ErrorJob,
   ErrorQualification,
 } from '../../common/constants/errors';
+import { TOKEN_ADDRESSES } from '../../common/constants/tokens';
+import { CronJobType } from '../../common/enums/cron-job';
 import {
+  EscrowFundToken,
   JobRequestType,
   JobStatus,
-  JobCaptchaMode,
-  JobCaptchaRequestType,
-  JobCaptchaShapeType,
-  EscrowFundToken,
 } from '../../common/enums/job';
 import { FiatCurrency } from '../../common/enums/payment';
+import { EventType, OracleType } from '../../common/enums/webhook';
+import { ControlledError } from '../../common/errors/controlled';
+import { PageDto } from '../../common/pagination/pagination.dto';
 import { parseUrl } from '../../common/utils';
-import { add, div, mul, max } from '../../common/utils/decimal';
-import { PaymentService } from '../payment/payment.service';
-import { Web3Service } from '../web3/web3.service';
+import { add, div, max, mul } from '../../common/utils/decimal';
+import { getTokenDecimals } from '../../common/utils/tokens';
+import { CronJobRepository } from '../cron-job/cron-job.repository';
 import {
   CvatManifestDto,
+  FortuneManifestDto,
+  HCaptchaManifestDto,
+} from '../manifest/manifest.dto';
+import { ManifestService } from '../manifest/manifest.service';
+import { PaymentService } from '../payment/payment.service';
+import { QualificationService } from '../qualification/qualification.service';
+import { RateService } from '../rate/rate.service';
+import { RoutingProtocolService } from '../routing-protocol/routing-protocol.service';
+import { StorageService } from '../storage/storage.service';
+import { UserEntity } from '../user/user.entity';
+import { Web3Service } from '../web3/web3.service';
+import { WebhookDataDto } from '../webhook/webhook.dto';
+import { WebhookEntity } from '../webhook/webhook.entity';
+import { WebhookRepository } from '../webhook/webhook.repository';
+import { WhitelistService } from '../whitelist/whitelist.service';
+import {
+  CreateJob,
   EscrowCancelDto,
   FortuneFinalResultDto,
-  FortuneManifestDto,
+  GetJobsDto,
+  JobAudinoDto,
+  JobCaptchaDto,
   JobCvatDto,
   JobDetailsDto,
   JobFortuneDto,
   JobListDto,
-  HCaptchaManifestDto,
-  JobCaptchaAdvancedDto,
-  JobCaptchaDto,
-  RestrictedAudience,
-  CreateJob,
   JobQuickLaunchDto,
-  CvatDataDto,
-  StorageDataDto,
-  GetJobsDto,
-  JobAudinoDto,
-  AudinoManifestDto,
 } from './job.dto';
 import { JobEntity } from './job.entity';
+import { EscrowAction, RequestAction } from './job.interface';
 import { JobRepository } from './job.repository';
-import {
-  CANCEL_JOB_STATUSES,
-  HCAPTCHA_BOUNDING_BOX_MAX_POINTS,
-  HCAPTCHA_BOUNDING_BOX_MIN_POINTS,
-  HCAPTCHA_IMMO_MAX_LENGTH,
-  HCAPTCHA_IMMO_MIN_LENGTH,
-  HCAPTCHA_LANDMARK_MAX_POINTS,
-  HCAPTCHA_LANDMARK_MIN_POINTS,
-  HCAPTCHA_MAX_SHAPES_PER_IMAGE,
-  HCAPTCHA_MINIMUM_SELECTION_AREA_PER_SHAPE,
-  HCAPTCHA_MIN_SHAPES_PER_IMAGE,
-  HCAPTCHA_NOT_PRESENTED_LABEL,
-  HCAPTCHA_ORACLE_STAKE,
-  HCAPTCHA_POLYGON_MAX_POINTS,
-  HCAPTCHA_POLYGON_MIN_POINTS,
-} from '../../common/constants';
-import { EventType, OracleType } from '../../common/enums/webhook';
-import {
-  HMToken,
-  HMToken__factory,
-} from '@human-protocol/core/typechain-types';
-import Decimal from 'decimal.js';
-import { StorageService } from '../storage/storage.service';
-import {
-  generateBucketUrl,
-  listObjectsInBucket,
-} from '../../common/utils/storage';
-import { WebhookDataDto } from '../webhook/webhook.dto';
-import {
-  ManifestAction,
-  EscrowAction,
-  OracleAction,
-  OracleAddresses,
-  RequestAction,
-  CvatCalculateJobBounty,
-  CvatImageData,
-  CvatAnnotationData,
-  GenerateUrls,
-} from './job.interface';
-import { WebhookEntity } from '../webhook/webhook.entity';
-import { WebhookRepository } from '../webhook/webhook.repository';
-import { ControlledError } from '../../common/errors/controlled';
-import { RateService } from '../rate/rate.service';
-import { PageDto } from '../../common/pagination/pagination.dto';
-import { CronJobType } from '../../common/enums/cron-job';
-import { CronJobRepository } from '../cron-job/cron-job.repository';
-import { ModuleRef } from '@nestjs/core';
-import { QualificationService } from '../qualification/qualification.service';
-import { WhitelistService } from '../whitelist/whitelist.service';
-import { UserEntity } from '../user/user.entity';
-import { RoutingProtocolService } from '../routing-protocol/routing-protocol.service';
-import { TOKEN_ADDRESSES } from '../../common/constants/tokens';
-import { getTokenDecimals } from '../../common/utils/tokens';
-
 @Injectable()
 export class JobService {
   public readonly logger = new Logger(JobService.name);
@@ -134,17 +92,13 @@ export class JobService {
     private readonly webhookRepository: WebhookRepository,
     private readonly paymentService: PaymentService,
     private readonly serverConfigService: ServerConfigService,
-    private readonly authConfigService: AuthConfigService,
-    private readonly web3ConfigService: Web3ConfigService,
-    private readonly cvatConfigService: CvatConfigService,
-    private readonly pgpConfigService: PGPConfigService,
     private readonly routingProtocolService: RoutingProtocolService,
     private readonly storageService: StorageService,
     private readonly rateService: RateService,
     private readonly whitelistService: WhitelistService,
     private moduleRef: ModuleRef,
-    @Inject(Encryption) private readonly encryption: Encryption,
     private readonly qualificationService: QualificationService,
+    private readonly manifestService: ManifestService,
   ) {}
 
   onModuleInit() {
@@ -152,330 +106,10 @@ export class JobService {
       strict: false,
     });
   }
-
-  public async createCvatManifest(
-    dto: JobCvatDto,
-    requestType: JobRequestType,
-    tokenFundAmount: number,
-  ): Promise<CvatManifestDto> {
-    const { generateUrls } = this.createManifestActions[requestType];
-
-    const urls = generateUrls(dto.data, dto.groundTruth);
-
-    const jobBounty = await this.calculateJobBounty({
-      requestType,
-      fundAmount: tokenFundAmount,
-      urls,
-      nodesTotal: dto.labels[0]?.nodes?.length,
-    });
-
-    return {
-      data: {
-        data_url: urls.dataUrl.href,
-        ...(urls.pointsUrl && {
-          points_url: urls.pointsUrl?.href,
-        }),
-        ...(urls.boxesUrl && {
-          boxes_url: urls.boxesUrl?.href,
-        }),
-      },
-      annotation: {
-        labels: dto.labels,
-        description: dto.requesterDescription,
-        user_guide: dto.userGuide,
-        type: requestType,
-        job_size: this.cvatConfigService.jobSize,
-        ...(dto.qualifications && {
-          qualifications: dto.qualifications,
-        }),
-      },
-      validation: {
-        min_quality: dto.minQuality,
-        val_size: this.cvatConfigService.valSize,
-        gt_url: urls.gtUrl.href,
-      },
-      job_bounty: jobBounty,
-    };
-  }
-
-  public async createHCaptchaManifest(
-    jobDto: JobCaptchaDto,
-  ): Promise<HCaptchaManifestDto> {
-    const jobType = jobDto.annotations.typeOfJob;
-    const dataUrl = generateBucketUrl(jobDto.data, JobRequestType.HCAPTCHA);
-    const objectsInBucket = await listObjectsInBucket(dataUrl);
-
-    const commonManifestProperties = {
-      job_mode: JobCaptchaMode.BATCH,
-      requester_accuracy_target: jobDto.accuracyTarget,
-      request_config: {},
-      restricted_audience: this.buildHCaptchaRestrictedAudience(
-        jobDto.advanced,
-      ),
-      requester_max_repeats: jobDto.maxRequests,
-      requester_min_repeats: jobDto.minRequests,
-      requester_question: { en: jobDto.annotations.labelingPrompt },
-      job_total_tasks: objectsInBucket.length,
-      task_bid_price: jobDto.annotations.taskBidPrice,
-      taskdata_uri: await this.generateAndUploadTaskData(
-        dataUrl.href,
-        objectsInBucket,
-      ),
-      public_results: true,
-      oracle_stake: HCAPTCHA_ORACLE_STAKE,
-      repo_uri: this.web3ConfigService.hCaptchaReputationOracleURI,
-      ro_uri: this.web3ConfigService.hCaptchaRecordingOracleURI,
-      ...(jobDto.qualifications && {
-        qualifications: jobDto.qualifications,
-      }),
-    };
-
-    let groundTruthsData;
-    if (jobDto.annotations.groundTruths) {
-      groundTruthsData = await this.storageService.downloadJsonLikeData(
-        jobDto.annotations.groundTruths,
-      );
-    }
-
-    switch (jobType) {
-      case JobCaptchaShapeType.COMPARISON:
-        return {
-          ...commonManifestProperties,
-          request_type: JobCaptchaRequestType.IMAGE_LABEL_BINARY,
-          groundtruth_uri: jobDto.annotations.groundTruths,
-          requester_restricted_answer_set: {},
-          requester_question_example: jobDto.annotations.exampleImages || [],
-        };
-
-      case JobCaptchaShapeType.CATEGORIZATION:
-        return {
-          ...commonManifestProperties,
-          request_type: JobCaptchaRequestType.IMAGE_LABEL_MULTIPLE_CHOICE,
-          groundtruth_uri: jobDto.annotations.groundTruths,
-          requester_restricted_answer_set:
-            this.buildHCaptchaRestrictedAnswerSet(groundTruthsData),
-        };
-
-      case JobCaptchaShapeType.POLYGON:
-        if (!jobDto.annotations.label) {
-          throw new ControlledError(
-            ErrorJob.JobParamsValidationFailed,
-            HttpStatus.BAD_REQUEST,
-          );
-        }
-
-        const polygonManifest = {
-          ...commonManifestProperties,
-          request_type: JobCaptchaRequestType.IMAGE_LABEL_AREA_SELECT,
-          request_config: {
-            shape_type: JobCaptchaShapeType.POLYGON,
-            min_shapes_per_image: HCAPTCHA_MIN_SHAPES_PER_IMAGE,
-            max_shapes_per_image: HCAPTCHA_MAX_SHAPES_PER_IMAGE,
-            min_points: HCAPTCHA_POLYGON_MIN_POINTS,
-            max_points: HCAPTCHA_POLYGON_MAX_POINTS,
-            minimum_selection_area_per_shape:
-              HCAPTCHA_MINIMUM_SELECTION_AREA_PER_SHAPE,
-          },
-          groundtruth_uri: jobDto.annotations.groundTruths,
-          requester_restricted_answer_set: {
-            [jobDto.annotations.label!]: { en: jobDto.annotations.label },
-          },
-          requester_question_example: jobDto.annotations.exampleImages || [],
-        };
-
-        return polygonManifest;
-
-      case JobCaptchaShapeType.POINT:
-        if (!jobDto.annotations.label) {
-          throw new ControlledError(
-            ErrorJob.JobParamsValidationFailed,
-            HttpStatus.BAD_REQUEST,
-          );
-        }
-
-        const pointManifest = {
-          ...commonManifestProperties,
-          request_type: JobCaptchaRequestType.IMAGE_LABEL_AREA_SELECT,
-          request_config: {
-            shape_type: jobType,
-            min_shapes_per_image: HCAPTCHA_MIN_SHAPES_PER_IMAGE,
-            max_shapes_per_image: HCAPTCHA_MAX_SHAPES_PER_IMAGE,
-            min_points: HCAPTCHA_LANDMARK_MIN_POINTS,
-            max_points: HCAPTCHA_LANDMARK_MAX_POINTS,
-          },
-          groundtruth_uri: jobDto.annotations.groundTruths,
-          requester_restricted_answer_set: {
-            [jobDto.annotations.label!]: { en: jobDto.annotations.label },
-          },
-          requester_question_example: jobDto.annotations.exampleImages || [],
-        };
-
-        return pointManifest;
-      case JobCaptchaShapeType.BOUNDING_BOX:
-        if (!jobDto.annotations.label) {
-          throw new ControlledError(
-            ErrorJob.JobParamsValidationFailed,
-            HttpStatus.BAD_REQUEST,
-          );
-        }
-
-        const boundingBoxManifest = {
-          ...commonManifestProperties,
-          request_type: JobCaptchaRequestType.IMAGE_LABEL_AREA_SELECT,
-          request_config: {
-            shape_type: jobType,
-            min_shapes_per_image: HCAPTCHA_MIN_SHAPES_PER_IMAGE,
-            max_shapes_per_image: HCAPTCHA_MAX_SHAPES_PER_IMAGE,
-            min_points: HCAPTCHA_BOUNDING_BOX_MIN_POINTS,
-            max_points: HCAPTCHA_BOUNDING_BOX_MAX_POINTS,
-          },
-          groundtruth_uri: jobDto.annotations.groundTruths,
-          requester_restricted_answer_set: {
-            [jobDto.annotations.label!]: { en: jobDto.annotations.label },
-          },
-          requester_question_example: jobDto.annotations.exampleImages || [],
-        };
-
-        return boundingBoxManifest;
-      case JobCaptchaShapeType.IMMO:
-        if (!jobDto.annotations.label) {
-          throw new ControlledError(
-            ErrorJob.JobParamsValidationFailed,
-            HttpStatus.BAD_REQUEST,
-          );
-        }
-
-        const immoManifest = {
-          ...commonManifestProperties,
-          request_type: JobCaptchaRequestType.TEXT_FREEE_NTRY,
-          request_config: {
-            multiple_choice_max_choices: 1,
-            multiple_choice_min_choices: 1,
-            overlap_threshold: null,
-            answer_type: 'str',
-            max_length: HCAPTCHA_IMMO_MAX_LENGTH,
-            min_length: HCAPTCHA_IMMO_MIN_LENGTH,
-          },
-          requester_restricted_answer_set: {
-            [jobDto.annotations.label!]: { en: jobDto.annotations.label },
-          },
-          taskdata: [],
-        };
-
-        return immoManifest;
-
-      default:
-        throw new ControlledError(
-          ErrorJob.HCaptchaInvalidJobType,
-          HttpStatus.CONFLICT,
-        );
-    }
-  }
-
-  public async createAudinoManifest(
-    dto: JobAudinoDto,
-    requestType: JobRequestType,
-    tokenFundAmount: number,
-  ): Promise<any> {
-    const { generateUrls } = this.createManifestActions[requestType];
-    const urls = generateUrls(dto.data, dto.groundTruth);
-    const totalSegments = Math.ceil(
-      (dto.audioDuration * 1000) / dto.segmentDuration,
-    );
-
-    const jobBounty =
-      ethers.parseUnits(tokenFundAmount.toString(), 'ether') /
-      BigInt(totalSegments);
-
-    return {
-      annotation: {
-        description: dto.requesterDescription,
-        labels: dto.labels,
-        qualifications: dto.qualifications || [],
-        type: requestType,
-        user_guide: dto.userGuide,
-        segment_duration: dto.segmentDuration,
-      },
-      data: {
-        data_url: urls.dataUrl.href,
-      },
-      job_bounty: ethers.formatEther(jobBounty),
-      validation: {
-        gt_url: urls.gtUrl.href,
-        min_quality: dto.minQuality,
-      },
-    };
-  }
-
-  private buildHCaptchaRestrictedAudience(advanced: JobCaptchaAdvancedDto) {
-    const restrictedAudience: RestrictedAudience = {};
-
-    restrictedAudience.sitekey = [
-      {
-        [this.authConfigService.hCaptchaSiteKey]: {
-          score: 1,
-        },
-      },
-    ];
-
-    if (advanced.workerLanguage) {
-      restrictedAudience.lang = [{ [advanced.workerLanguage]: { score: 1 } }];
-    }
-
-    if (advanced.workerLocation) {
-      restrictedAudience.country = [
-        { [advanced.workerLocation]: { score: 1 } },
-      ];
-    }
-
-    if (advanced.targetBrowser) {
-      restrictedAudience.browser = [{ [advanced.targetBrowser]: { score: 1 } }];
-    }
-
-    return restrictedAudience;
-  }
-
-  private buildHCaptchaRestrictedAnswerSet(groundTruthsData: any) {
-    const maxElements = 3;
-    const outputObject: any = {};
-
-    let elementCount = 0;
-
-    for (const key of Object.keys(groundTruthsData)) {
-      if (elementCount >= maxElements) {
-        break;
-      }
-
-      const value = groundTruthsData[key][0][0];
-      outputObject[value] = { en: value, answer_example_uri: key };
-      elementCount++;
-    }
-
-    // Default case
-    outputObject['0'] = { en: HCAPTCHA_NOT_PRESENTED_LABEL };
-
-    return outputObject;
-  }
-
-  public async generateAndUploadTaskData(
-    dataUrl: string,
-    objectNames: string[],
-  ) {
-    const data = objectNames.map((objectName) => {
-      return {
-        datapoint_uri: `${dataUrl}/${objectName}`,
-        datapoint_hash: 'undefined-hash',
-        task_key: uuidv4(),
-      };
-    });
-
-    const { url } = await this.storageService.uploadJsonLikeData(data);
-    return url;
-  }
-
   private createJobSpecificActions: Record<JobRequestType, RequestAction> = {
     [JobRequestType.HCAPTCHA]: {
-      createManifest: (dto: JobCaptchaDto) => this.createHCaptchaManifest(dto),
+      createManifest: (dto: JobCaptchaDto) =>
+        this.manifestService.createHCaptchaManifest(dto),
     },
     [JobRequestType.FORTUNE]: {
       createManifest: async (
@@ -493,42 +127,48 @@ export class JobService {
         dto: JobCvatDto,
         requestType: JobRequestType,
         fundAmount: number,
-      ) => this.createCvatManifest(dto, requestType, fundAmount),
+      ) =>
+        this.manifestService.createCvatManifest(dto, requestType, fundAmount),
     },
     [JobRequestType.IMAGE_BOXES]: {
       createManifest: (
         dto: JobCvatDto,
         requestType: JobRequestType,
         fundAmount: number,
-      ) => this.createCvatManifest(dto, requestType, fundAmount),
+      ) =>
+        this.manifestService.createCvatManifest(dto, requestType, fundAmount),
     },
     [JobRequestType.IMAGE_POINTS]: {
       createManifest: (
         dto: JobCvatDto,
         requestType: JobRequestType,
         fundAmount: number,
-      ) => this.createCvatManifest(dto, requestType, fundAmount),
+      ) =>
+        this.manifestService.createCvatManifest(dto, requestType, fundAmount),
     },
     [JobRequestType.IMAGE_BOXES_FROM_POINTS]: {
       createManifest: (
         dto: JobCvatDto,
         requestType: JobRequestType,
         fundAmount: number,
-      ) => this.createCvatManifest(dto, requestType, fundAmount),
+      ) =>
+        this.manifestService.createCvatManifest(dto, requestType, fundAmount),
     },
     [JobRequestType.IMAGE_SKELETONS_FROM_BOXES]: {
       createManifest: (
         dto: JobCvatDto,
         requestType: JobRequestType,
         fundAmount: number,
-      ) => this.createCvatManifest(dto, requestType, fundAmount),
+      ) =>
+        this.manifestService.createCvatManifest(dto, requestType, fundAmount),
     },
     [JobRequestType.AUDIO_TRANSCRIPTION]: {
       createManifest: (
         dto: JobAudinoDto,
         requestType: JobRequestType,
         fundAmount: number,
-      ) => this.createAudinoManifest(dto, requestType, fundAmount),
+      ) =>
+        this.manifestService.createAudinoManifest(dto, requestType, fundAmount),
     },
   };
 
@@ -559,279 +199,13 @@ export class JobService {
     },
   };
 
-  private createManifestActions: Record<JobRequestType, ManifestAction> = {
-    [JobRequestType.HCAPTCHA]: {
-      getElementsCount: async () => 0,
-      generateUrls: () => ({ dataUrl: new URL(''), gtUrl: new URL('') }),
-    },
-    [JobRequestType.FORTUNE]: {
-      getElementsCount: async () => 0,
-      generateUrls: () => ({ dataUrl: new URL(''), gtUrl: new URL('') }),
-    },
-    [JobRequestType.IMAGE_POLYGONS]: {
-      getElementsCount: async (urls: GenerateUrls) => {
-        const gt = (await this.storageService.downloadJsonLikeData(
-          `${urls.gtUrl.protocol}//${urls.gtUrl.host}${urls.gtUrl.pathname}`,
-        )) as any;
-        if (!gt || !gt.images || gt.images.length === 0)
-          throw new ControlledError(
-            ErrorJob.GroundThuthValidationFailed,
-            HttpStatus.BAD_REQUEST,
-          );
-
-        const data = await listObjectsInBucket(urls.dataUrl);
-        if (!data || data.length === 0 || !data[0])
-          throw new ControlledError(
-            ErrorJob.DatasetValidationFailed,
-            HttpStatus.BAD_REQUEST,
-          );
-
-        await this.checkImageConsistency(gt.images, data);
-
-        return data.length - gt.images.length;
-      },
-      generateUrls: (
-        data: CvatDataDto,
-        groundTruth: StorageDataDto,
-      ): GenerateUrls => {
-        const requestType = JobRequestType.IMAGE_POLYGONS;
-        return {
-          dataUrl: generateBucketUrl(data.dataset, requestType),
-          gtUrl: generateBucketUrl(groundTruth, requestType),
-        };
-      },
-    },
-    [JobRequestType.IMAGE_BOXES]: {
-      getElementsCount: async (urls: GenerateUrls) => {
-        const gt = (await this.storageService.downloadJsonLikeData(
-          `${urls.gtUrl.protocol}//${urls.gtUrl.host}${urls.gtUrl.pathname}`,
-        )) as any;
-        if (!gt || !gt.images || gt.images.length === 0)
-          throw new ControlledError(
-            ErrorJob.GroundThuthValidationFailed,
-            HttpStatus.BAD_REQUEST,
-          );
-
-        const data = await listObjectsInBucket(urls.dataUrl);
-        if (!data || data.length === 0 || !data[0])
-          throw new ControlledError(
-            ErrorJob.DatasetValidationFailed,
-            HttpStatus.BAD_REQUEST,
-          );
-
-        await this.checkImageConsistency(gt.images, data);
-
-        return data.length - gt.images.length;
-      },
-      generateUrls: (
-        data: CvatDataDto,
-        groundTruth: StorageDataDto,
-      ): GenerateUrls => {
-        const requestType = JobRequestType.IMAGE_BOXES;
-        return {
-          dataUrl: generateBucketUrl(data.dataset, requestType),
-          gtUrl: generateBucketUrl(groundTruth, requestType),
-        };
-      },
-    },
-    [JobRequestType.IMAGE_POINTS]: {
-      getElementsCount: async (urls: GenerateUrls) => {
-        const gt = (await this.storageService.downloadJsonLikeData(
-          `${urls.gtUrl.protocol}//${urls.gtUrl.host}${urls.gtUrl.pathname}`,
-        )) as any;
-        if (!gt || !gt.images || gt.images.length === 0)
-          throw new ControlledError(
-            ErrorJob.GroundThuthValidationFailed,
-            HttpStatus.BAD_REQUEST,
-          );
-
-        const data = await listObjectsInBucket(urls.dataUrl);
-        if (!data || data.length === 0 || !data[0])
-          throw new ControlledError(
-            ErrorJob.DatasetValidationFailed,
-            HttpStatus.BAD_REQUEST,
-          );
-
-        await this.checkImageConsistency(gt.images, data);
-
-        return data.length - gt.images.length;
-      },
-      generateUrls: (
-        data: CvatDataDto,
-        groundTruth: StorageDataDto,
-      ): GenerateUrls => {
-        const requestType = JobRequestType.IMAGE_POINTS;
-
-        return {
-          dataUrl: generateBucketUrl(data.dataset, requestType),
-          gtUrl: generateBucketUrl(groundTruth, requestType),
-        };
-      },
-    },
-    [JobRequestType.IMAGE_BOXES_FROM_POINTS]: {
-      getElementsCount: async (urls: GenerateUrls) =>
-        this.getCvatElementsCount(urls.gtUrl, urls.pointsUrl!),
-      generateUrls: (
-        data: CvatDataDto,
-        groundTruth: StorageDataDto,
-      ): GenerateUrls => {
-        if (!data.points) {
-          throw new ControlledError(ErrorJob.DataNotExist, HttpStatus.CONFLICT);
-        }
-
-        const requestType = JobRequestType.IMAGE_BOXES_FROM_POINTS;
-
-        return {
-          dataUrl: generateBucketUrl(data.dataset, requestType),
-          gtUrl: generateBucketUrl(groundTruth, requestType),
-          pointsUrl: generateBucketUrl(data.points, requestType),
-        };
-      },
-    },
-    [JobRequestType.IMAGE_SKELETONS_FROM_BOXES]: {
-      getElementsCount: async (urls: GenerateUrls) =>
-        this.getCvatElementsCount(urls.gtUrl, urls.boxesUrl!),
-      generateUrls: (
-        data: CvatDataDto,
-        groundTruth: StorageDataDto,
-      ): GenerateUrls => {
-        if (!data.boxes) {
-          throw new ControlledError(ErrorJob.DataNotExist, HttpStatus.CONFLICT);
-        }
-
-        const requestType = JobRequestType.IMAGE_SKELETONS_FROM_BOXES;
-
-        return {
-          dataUrl: generateBucketUrl(data.dataset, requestType),
-          gtUrl: generateBucketUrl(groundTruth, requestType),
-          boxesUrl: generateBucketUrl(data.boxes, requestType),
-        };
-      },
-    },
-    [JobRequestType.AUDIO_TRANSCRIPTION]: {
-      getElementsCount: async () => 0,
-      generateUrls: (
-        data: CvatDataDto,
-        groundTruth: StorageDataDto,
-      ): GenerateUrls => {
-        const requestType = JobRequestType.AUDIO_TRANSCRIPTION;
-
-        return {
-          dataUrl: generateBucketUrl(data.dataset, requestType),
-          gtUrl: generateBucketUrl(groundTruth, requestType),
-        };
-      },
-    },
-  };
-
-  private getOraclesSpecificActions: Record<JobRequestType, OracleAction> = {
-    [JobRequestType.HCAPTCHA]: {
-      getOracleAddresses: (): OracleAddresses => {
-        const exchangeOracle = this.web3ConfigService.hCaptchaOracleAddress;
-        const recordingOracle = this.web3ConfigService.hCaptchaOracleAddress;
-        const reputationOracle = this.web3ConfigService.hCaptchaOracleAddress;
-
-        return { exchangeOracle, recordingOracle, reputationOracle };
-      },
-    },
-    [JobRequestType.FORTUNE]: {
-      getOracleAddresses: (): OracleAddresses => {
-        const exchangeOracle =
-          this.web3ConfigService.fortuneExchangeOracleAddress;
-        const recordingOracle =
-          this.web3ConfigService.fortuneRecordingOracleAddress;
-        const reputationOracle = this.web3ConfigService.reputationOracleAddress;
-
-        return { exchangeOracle, recordingOracle, reputationOracle };
-      },
-    },
-    [JobRequestType.IMAGE_POLYGONS]: {
-      getOracleAddresses: (): OracleAddresses => {
-        const exchangeOracle = this.web3ConfigService.cvatExchangeOracleAddress;
-        const recordingOracle =
-          this.web3ConfigService.cvatRecordingOracleAddress;
-        const reputationOracle = this.web3ConfigService.reputationOracleAddress;
-
-        return { exchangeOracle, recordingOracle, reputationOracle };
-      },
-    },
-    [JobRequestType.IMAGE_BOXES]: {
-      getOracleAddresses: (): OracleAddresses => {
-        const exchangeOracle = this.web3ConfigService.cvatExchangeOracleAddress;
-        const recordingOracle =
-          this.web3ConfigService.cvatRecordingOracleAddress;
-        const reputationOracle = this.web3ConfigService.reputationOracleAddress;
-
-        return { exchangeOracle, recordingOracle, reputationOracle };
-      },
-    },
-    [JobRequestType.IMAGE_POINTS]: {
-      getOracleAddresses: (): OracleAddresses => {
-        const exchangeOracle = this.web3ConfigService.cvatExchangeOracleAddress;
-        const recordingOracle =
-          this.web3ConfigService.cvatRecordingOracleAddress;
-        const reputationOracle = this.web3ConfigService.reputationOracleAddress;
-
-        return { exchangeOracle, recordingOracle, reputationOracle };
-      },
-    },
-    [JobRequestType.IMAGE_BOXES_FROM_POINTS]: {
-      getOracleAddresses: (): OracleAddresses => {
-        const exchangeOracle = this.web3ConfigService.cvatExchangeOracleAddress;
-        const recordingOracle =
-          this.web3ConfigService.cvatRecordingOracleAddress;
-        const reputationOracle = this.web3ConfigService.reputationOracleAddress;
-
-        return { exchangeOracle, recordingOracle, reputationOracle };
-      },
-    },
-    [JobRequestType.IMAGE_SKELETONS_FROM_BOXES]: {
-      getOracleAddresses: (): OracleAddresses => {
-        const exchangeOracle = this.web3ConfigService.cvatExchangeOracleAddress;
-        const recordingOracle =
-          this.web3ConfigService.cvatRecordingOracleAddress;
-        const reputationOracle = this.web3ConfigService.reputationOracleAddress;
-
-        return { exchangeOracle, recordingOracle, reputationOracle };
-      },
-    },
-    [JobRequestType.AUDIO_TRANSCRIPTION]: {
-      getOracleAddresses: (): OracleAddresses => {
-        return {
-          exchangeOracle: this.web3ConfigService.audinoExchangeOracleAddress,
-          recordingOracle: this.web3ConfigService.audinoRecordingOracleAddress,
-          reputationOracle: this.web3ConfigService.reputationOracleAddress,
-        };
-      },
-    },
-  };
-
-  private async checkImageConsistency(
-    gtImages: any[],
-    dataFiles: string[],
-  ): Promise<void> {
-    const gtFileNames = gtImages.map((image: any) => image.file_name);
-    const baseFileNames = dataFiles.map((fileName) =>
-      fileName.split('/').pop(),
-    );
-    const missingFileNames = gtFileNames.filter(
-      (fileName: any) => !baseFileNames.includes(fileName),
-    );
-
-    if (missingFileNames.length !== 0) {
-      throw new ControlledError(
-        ErrorJob.ImageConsistency,
-        HttpStatus.BAD_REQUEST,
-      );
-    }
-  }
-
   public async createJob(
     user: UserEntity,
     requestType: JobRequestType,
     dto: CreateJob,
   ): Promise<number> {
     let { chainId, reputationOracle, exchangeOracle, recordingOracle } = dto;
+
     if (!Object.values(JobRequestType).includes(requestType)) {
       throw new ControlledError(
         ErrorJob.InvalidRequestType,
@@ -976,10 +350,10 @@ export class JobService {
         fundTokenAmount,
       );
 
-      const { url, hash } = await this.uploadManifest(
-        requestType,
+      const { url, hash } = await this.manifestService.uploadManifest(
         chainId,
         manifestOrigin,
+        [exchangeOracle, reputationOracle, recordingOracle],
       );
 
       jobEntity.manifestUrl = url;
@@ -1019,73 +393,6 @@ export class JobService {
     return jobEntity.id;
   }
 
-  public async getCvatElementsCount(gtUrl: URL, dataUrl: URL): Promise<number> {
-    const data = (await this.storageService.downloadJsonLikeData(
-      dataUrl.href,
-    )) as any;
-    const gt = (await this.storageService.downloadJsonLikeData(
-      gtUrl.href,
-    )) as any;
-
-    if (!gt || !gt.images || gt.images.length === 0) {
-      throw new ControlledError(
-        ErrorJob.GroundThuthValidationFailed,
-        HttpStatus.BAD_REQUEST,
-      );
-    }
-
-    let gtEntries = 0;
-
-    gt.images.forEach((gtImage: CvatImageData) => {
-      const { id } = data.images.find(
-        (dataImage: CvatImageData) => dataImage.file_name === gtImage.file_name,
-      );
-
-      if (id) {
-        const matchingAnnotations = data.annotations.filter(
-          (dataAnnotation: CvatAnnotationData) =>
-            dataAnnotation.image_id === id,
-        );
-        gtEntries += matchingAnnotations.length;
-      }
-    });
-
-    return data.annotations.length - gtEntries;
-  }
-
-  public async calculateJobBounty(
-    params: CvatCalculateJobBounty,
-  ): Promise<string> {
-    const { requestType, fundAmount, urls, nodesTotal } = params;
-
-    const { getElementsCount } = this.createManifestActions[requestType];
-    const elementsCount = await getElementsCount(urls);
-
-    let jobSize = Number(this.cvatConfigService.jobSize);
-
-    if (requestType === JobRequestType.IMAGE_SKELETONS_FROM_BOXES) {
-      const jobSizeMultiplier = Number(
-        this.cvatConfigService.skeletonsJobSizeMultiplier,
-      );
-      jobSize *= jobSizeMultiplier;
-    }
-
-    let totalJobs: number;
-
-    // For each skeleton node CVAT creates a separate project thus increasing amount of jobs
-    if (
-      requestType === JobRequestType.IMAGE_SKELETONS_FROM_BOXES &&
-      nodesTotal
-    ) {
-      totalJobs = Math.ceil(elementsCount / jobSize) * nodesTotal;
-    } else {
-      totalJobs = Math.ceil(elementsCount / jobSize);
-    }
-    const jobBounty =
-      ethers.parseUnits(fundAmount.toString(), 'ether') / BigInt(totalJobs);
-    return ethers.formatEther(jobBounty);
-  }
-
   public async createEscrow(jobEntity: JobEntity): Promise<JobEntity> {
     const { getTrustedHandlers } =
       this.createEscrowSpecificActions[jobEntity.requestType];
@@ -1117,33 +424,24 @@ export class JobService {
   }
 
   public async setupEscrow(jobEntity: JobEntity): Promise<JobEntity> {
-    const { getOracleAddresses } =
-      this.getOraclesSpecificActions[jobEntity.requestType];
-
     const signer = this.web3Service.getSigner(jobEntity.chainId);
 
     const escrowClient = await EscrowClient.build(signer);
 
-    const manifest = (await this.storageService.downloadJsonLikeData(
-      jobEntity.manifestUrl,
-    )) as any;
-
-    await this.validateManifest(jobEntity.requestType, manifest);
-
-    const oracleAddresses = getOracleAddresses();
-
     const escrowConfig = {
-      ...oracleAddresses,
+      recordingOracle: jobEntity.recordingOracle,
       recordingOracleFee: await this.getOracleFee(
-        oracleAddresses.recordingOracle,
+        jobEntity.recordingOracle,
         jobEntity.chainId,
       ),
+      reputationOracle: jobEntity.reputationOracle,
       reputationOracleFee: await this.getOracleFee(
-        oracleAddresses.reputationOracle,
+        jobEntity.reputationOracle,
         jobEntity.chainId,
       ),
+      exchangeOracle: jobEntity.exchangeOracle,
       exchangeOracleFee: await this.getOracleFee(
-        oracleAddresses.exchangeOracle,
+        jobEntity.exchangeOracle,
         jobEntity.chainId,
       ),
       manifestUrl: jobEntity.manifestUrl,
@@ -1289,81 +587,6 @@ export class JobService {
 
     jobEntity.retriesCount = 0;
     await this.jobRepository.updateOne(jobEntity);
-  }
-
-  public async uploadManifest(
-    requestType: JobRequestType,
-    chainId: ChainId,
-    data: any,
-  ): Promise<any> {
-    let manifestFile = data;
-
-    if (this.pgpConfigService.encrypt) {
-      const { getOracleAddresses } =
-        this.getOraclesSpecificActions[requestType];
-
-      const signer = this.web3Service.getSigner(chainId);
-      const publicKeys: string[] = [
-        await KVStoreUtils.getPublicKey(chainId, signer.address),
-      ];
-      const oracleAddresses = getOracleAddresses();
-      for (const address of Object.values(oracleAddresses)) {
-        const publicKey = await KVStoreUtils.getPublicKey(chainId, address);
-        if (publicKey) publicKeys.push(publicKey);
-      }
-      const encryptedManifest = await this.encryption.signAndEncrypt(
-        JSON.stringify(data),
-        publicKeys,
-      );
-      manifestFile = encryptedManifest;
-    }
-
-    const uploadedFile =
-      await this.storageService.uploadJsonLikeData(manifestFile);
-
-    if (!uploadedFile) {
-      throw new ControlledError(
-        ErrorBucket.UnableSaveFile,
-        HttpStatus.BAD_REQUEST,
-      );
-    }
-
-    return uploadedFile;
-  }
-
-  private async validateManifest(
-    requestType: JobRequestType,
-    manifest: FortuneManifestDto | CvatManifestDto | HCaptchaManifestDto,
-  ): Promise<boolean> {
-    let dtoCheck;
-
-    if (requestType === JobRequestType.FORTUNE) {
-      dtoCheck = new FortuneManifestDto();
-    } else if (requestType === JobRequestType.HCAPTCHA) {
-      return true;
-      dtoCheck = new HCaptchaManifestDto();
-    } else if (requestType === JobRequestType.AUDIO_TRANSCRIPTION) {
-      dtoCheck = new AudinoManifestDto();
-    } else {
-      dtoCheck = new CvatManifestDto();
-    }
-
-    Object.assign(dtoCheck, manifest);
-
-    const validationErrors: ValidationError[] = await validate(dtoCheck);
-    if (validationErrors.length > 0) {
-      this.logger.log(
-        ErrorJob.ManifestValidationFailed,
-        JobService.name,
-        validationErrors,
-      );
-      throw new ControlledError(
-        ErrorJob.ManifestValidationFailed,
-        HttpStatus.NOT_FOUND,
-      );
-    }
-
-    return true;
   }
 
   public async getJobsByStatus(
