+++ packages/apps/job-launcher/server/src/modules/manifest/manifest.module.ts
@@ -1,0 +1,32 @@
+import { Module } from '@nestjs/common';
+import { ManifestService } from './manifest.service';
+import { StorageModule } from '../storage/storage.module';
+import { Web3Module } from '../web3/web3.module';
+import { EncryptionModule } from '../encryption/encryption.module';
+import { RoutingProtocolModule } from '../routing-protocol/routing-protocol.module';
+import { RateModule } from '../rate/rate.module';
+import { QualificationModule } from '../qualification/qualification.module';
+import { AuthConfigService } from '../../common/config/auth-config.service';
+import { CvatConfigService } from '../../common/config/cvat-config.service';
+import { Web3ConfigService } from '../../common/config/web3-config.service';
+import { PGPConfigService } from '../../common/config/pgp-config.service';
+
+@Module({
+  imports: [
+    StorageModule,
+    Web3Module,
+    EncryptionModule,
+    RoutingProtocolModule,
+    RateModule,
+    QualificationModule,
+  ],
+  providers: [
+    ManifestService,
+    AuthConfigService,
+    CvatConfigService,
+    Web3ConfigService,
+    PGPConfigService,
+  ],
+  exports: [ManifestService],
+})
+export class ManifestModule {}
